requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.7.1
black==19.3b0
cffi==1.12.2
Click==7.0
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
lxml==4.3.3
parsel==1.5.1
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
queuelib==1.5.0
Scrapy==1.6.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9
toml==0.10.0
Twisted==18.9.0
w3lib==1.20.0
zope.interface==4.6.0
